Executes $apply() only if doesn't have another in execution (blur)
Some components, like datepickers or combo boxes or autocompletes,
has $apply() in her codes to "apply" something. Like this, sometimes
"angular-validation" needs apply something too.

When "angular-validation" applies something to some component and
that component executes $apply() before the validation, an error is
displayed in console: "$apply already in progress".

To correct this, just verify if $apply() is in progress.
If it isn't in progress execute him. If it is in progress,
doesn't execute him.